feat: PostItem compoenent for rendering posts lists
